Style Install App
I added some CSS stylings to the application. I also fixed some of the
naming errors (a few of the classes in index.html have spaces in the
names - those need to be replaced with _ or shortened.)

I will continue to work on this and tweak it for functionality similar
to the google sheets deal I've been working on.

My end goal for this project is for salespeople to be able to schedule
installers. In order to do that, they will need to see the installer's
availability.

I also want the salespeople to be able to create work orders from this
app.

A future update maybe could have the installers use it to see what their
jobs for us currently are.