Style Installation App with FlexBox
The install app is starting to take shape - I want to have a menu bar
across the top and the header immediately below that.
I had it looking the way I wanted to originally with drop-downs; I may
result to styling it back that way. The page will eventually be
responsive with JavaScript - if installer, it'll send you to one page,
if salesperson, it'll send you to sales_landing, if office,
office_landing. Etc.