Update App Layout
I decided to create a signup page, which will only appear if the user
has not signed up. Depending on if they sign up for "sales", "installer"
or "office" - that will determine what landing page they go to.
I added styling for a menu bar, though I'm not happy with that yet.
The menu bar will have different options on different pages - it
doesn't need to be on the signup page per se, but it is good to have it
there so I can make changes as needed and then copy/paste it to wherever
it needs to go.